backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
seaborn
pydantic
pydantic-settings
google-generativeai
langchain
langchain-google-genai
python-dotenv
aiofiles
joblib
